requirements.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.2

Pillow>=10.0
pyzbar>=0.1.9
